@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "uploadgram"
-version = "0.3.4"
+version = "0.4.0"
 description ="Upload files to Telegram upto 4 GiB, from the Terminal"
 authors = ["Shrimadhav U K <uploADGRam@shrimadhavUK.me>"]
 homepage = "https://github.com/SpEcHiDe/UploadGram"
@@ -17,27 +17,28 @@ classifiers = [
     "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
     "Programming Language :: Python",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
-    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
     "Programming Language :: Python :: Implementation :: CPython",
     "Programming Language :: Python :: Implementation :: PyPy",
     "Topic :: Software Development",
 ]
 
 [tool.poetry.dependencies]
-python = ">=3.7.7,<4"
-python-dotenv = "0.10"
-pyrotgfork = "2.1.33.8"
-TgCrypto = "1.2.5"
-hachoir = "3.1.1"
-tqdm = "4.62.3"
+python = "^3.10"
+python-dotenv = "^1.0.0"
+pyrogram = "^2.0.106"
+TgCrypto = "^1.2.5"
+hachoir = "^3.3.0"
+tqdm = "^4.66.0"
 
 [tool.poetry.dev-dependencies]
-pytest = "^5.2"
-isort = "^5.7.0"
-black = "^20.8b1"
-autopep8 = "^1.5.7"
-autoflake = "^1.4"
+pytest = "^7.0"
+isort = "^5.12.0"
+black = "^24.0"
+autopep8 = "^2.0.0"
+autoflake = "^2.0.0"
 
 [tool.isort]
 profile = "black"
